requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118

torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0

numpy>=1.21.0
opencv-python>=4.5.3
Pillow>=8.3.1
rembg>=2.0.30
trimesh>=3.9.35
pyrender>=0.1.45
matplotlib>=3.4.3
tqdm>=4.62.2
onnxruntime>=1.15.0
git+https://github.com/openai/shap-e.git
ninja
pyyaml>=6.0
ipywidgets>=8.0.0
